Move random and color functions to separate files
